refactor: 템플릿 부분을`TimetableSharingTemplate` 컴포넌트로 분리
